Python/Sync: Move config, exceptions, routes, proto codec to the shared folder
