+++ python/python/glide/__init__.py
@@ -1,60 +1,9 @@
 # Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0
 
-from glide.config import (
-    AdvancedGlideClientConfiguration,
-    AdvancedGlideClusterClientConfiguration,
-    BackoffStrategy,
-    GlideClientConfiguration,
-    GlideClusterClientConfiguration,
-    NodeAddress,
-    PeriodicChecksManualInterval,
-    PeriodicChecksStatus,
-    ProtocolVersion,
-    ReadFrom,
-    ServerCredentials,
-)
-from glide.constants import (
-    OK,
-    TOK,
-    FtAggregateResponse,
-    FtInfoResponse,
-    FtProfileResponse,
-    FtSearchResponse,
-    TClusterResponse,
-    TEncodable,
-    TFunctionListResponse,
-    TFunctionStatsFullResponse,
-    TFunctionStatsSingleNodeResponse,
-    TJsonResponse,
-    TJsonUniversalResponse,
-    TResult,
-    TSingleNodeRoute,
-    TXInfoStreamFullResponse,
-    TXInfoStreamResponse,
-)
-from glide.exceptions import (
-    ClosingError,
-    ConfigurationError,
-    ConnectionError,
-    ExecAbortError,
-    GlideError,
-    RequestError,
-    TimeoutError,
-)
 from glide.glide import ClusterScanCursor, Script
 from glide.glide_client import GlideClient, GlideClusterClient, TGlideClient
 from glide.logger import Level as LogLevel
 from glide.logger import Logger
-from glide.routes import (
-    AllNodes,
-    AllPrimaries,
-    ByAddressRoute,
-    RandomNode,
-    Route,
-    SlotIdRoute,
-    SlotKeyRoute,
-    SlotType,
-)
 from glide.shared.commands.batch import (
     Batch,
     ClusterBatch,
@@ -167,6 +116,57 @@ from glide.shared.commands.stream import (
     StreamTrimOptions,
     TrimByMaxLen,
     TrimByMinId,
+)
+from glide.shared.config import (
+    AdvancedGlideClientConfiguration,
+    AdvancedGlideClusterClientConfiguration,
+    BackoffStrategy,
+    GlideClientConfiguration,
+    GlideClusterClientConfiguration,
+    NodeAddress,
+    PeriodicChecksManualInterval,
+    PeriodicChecksStatus,
+    ProtocolVersion,
+    ReadFrom,
+    ServerCredentials,
+)
+from glide.shared.constants import (
+    OK,
+    TOK,
+    FtAggregateResponse,
+    FtInfoResponse,
+    FtProfileResponse,
+    FtSearchResponse,
+    TClusterResponse,
+    TEncodable,
+    TFunctionListResponse,
+    TFunctionStatsFullResponse,
+    TFunctionStatsSingleNodeResponse,
+    TJsonResponse,
+    TJsonUniversalResponse,
+    TResult,
+    TSingleNodeRoute,
+    TXInfoStreamFullResponse,
+    TXInfoStreamResponse,
+)
+from glide.shared.exceptions import (
+    ClosingError,
+    ConfigurationError,
+    ConnectionError,
+    ExecAbortError,
+    GlideError,
+    RequestError,
+    TimeoutError,
+)
+from glide.shared.routes import (
+    AllNodes,
+    AllPrimaries,
+    ByAddressRoute,
+    RandomNode,
+    Route,
+    SlotIdRoute,
+    SlotKeyRoute,
+    SlotType,
 )
 
 __all__ = [
