+++ python/python/glide/glide_client.py
@@ -19,14 +19,6 @@ import anyio
 import sniffio
 from anyio import to_thread
 
-from glide.config import BaseClientConfiguration, ServerCredentials
-from glide.constants import DEFAULT_READ_BYTES_SIZE, OK, TEncodable, TRequest, TResult
-from glide.exceptions import (
-    ClosingError,
-    ConfigurationError,
-    ConnectionError,
-    get_request_error_class,
-)
 from glide.glide import (
     DEFAULT_TIMEOUT_IN_MILLISECONDS,
     MAX_REQUEST_ARGS_LEN,
@@ -41,13 +33,27 @@ from glide.logger import Logger as ClientLogger
 from glide.protobuf.command_request_pb2 import Command, CommandRequest, RequestType
 from glide.protobuf.connection_request_pb2 import ConnectionRequest
 from glide.protobuf.response_pb2 import Response
-from glide.protobuf_codec import PartialMessageException, ProtobufCodec
-from glide.routes import Route, set_protobuf_route
 from glide.shared.commands.async_commands.cluster_commands import ClusterCommands
 from glide.shared.commands.async_commands.core import CoreCommands
 from glide.shared.commands.async_commands.standalone_commands import StandaloneCommands
 from glide.shared.commands.command_args import ObjectType
 from glide.shared.commands.core_options import PubSubMsg
+from glide.shared.config import BaseClientConfiguration, ServerCredentials
+from glide.shared.constants import (
+    DEFAULT_READ_BYTES_SIZE,
+    OK,
+    TEncodable,
+    TRequest,
+    TResult,
+)
+from glide.shared.exceptions import (
+    ClosingError,
+    ConfigurationError,
+    ConnectionError,
+    get_request_error_class,
+)
+from glide.shared.protobuf_codec import PartialMessageException, ProtobufCodec
+from glide.shared.routes import Route, set_protobuf_route
 
 if sys.version_info >= (3, 11):
     from typing import Self
