+++ python/python/glide/shared/commands/async_commands/standalone_commands.py
@@ -4,13 +4,6 @@ from __future__ import annotations
 
 from typing import Dict, List, Mapping, Optional, Union, cast
 
-from glide.constants import (
-    TOK,
-    TEncodable,
-    TFunctionListResponse,
-    TFunctionStatsFullResponse,
-    TResult,
-)
 from glide.glide import Script
 from glide.protobuf.command_request_pb2 import RequestType
 from glide.shared.commands.async_commands import CoreCommands
@@ -20,6 +13,13 @@ from glide.shared.commands.core_options import (
     FlushMode,
     FunctionRestorePolicy,
     InfoSection,
+)
+from glide.shared.constants import (
+    TOK,
+    TEncodable,
+    TFunctionListResponse,
+    TFunctionStatsFullResponse,
+    TResult,
 )
