+++ python/python/glide/shared/commands/server_modules/glide_json.py
@@ -19,9 +19,14 @@ Examples:
 """
 from typing import List, Optional, Union, cast
 
-from glide.constants import TOK, TEncodable, TJsonResponse, TJsonUniversalResponse
 from glide.glide_client import TGlideClient
 from glide.shared.commands.core_options import ConditionalChange
+from glide.shared.constants import (
+    TOK,
+    TEncodable,
+    TJsonResponse,
+    TJsonUniversalResponse,
+)
 
 
 class JsonGetOptions:
