+++ python/python/glide/shared/commands/sorted_set.py
@@ -3,8 +3,8 @@
 from enum import Enum
 from typing import List, Optional, Tuple, Union, cast
 
-from glide.constants import TEncodable
 from glide.shared.commands.command_args import Limit, OrderBy
+from glide.shared.constants import TEncodable
 
 
 class InfBound(Enum):
