+++ python/python/glide/shared/commands/stream.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from typing import List, Optional, Union
 
-from glide.constants import TEncodable
+from glide.shared.constants import TEncodable
 
 
 class StreamTrimOptions(ABC):
