+++ python/python/glide/shared/commands/sync_commands/cluster_commands.py
@@ -4,7 +4,14 @@ from __future__ import annotations
 
 from typing import Dict, List, Mapping, Optional, cast
 
-from glide.constants import (
+from glide.protobuf.command_request_pb2 import RequestType
+from glide.shared.commands.core_options import (
+    FlushMode,
+    FunctionRestorePolicy,
+    InfoSection,
+)
+from glide.shared.commands.sync_commands.core import CoreCommands
+from glide.shared.constants import (
     TOK,
     TClusterResponse,
     TEncodable,
@@ -12,14 +19,7 @@ from glide.constants import (
     TFunctionStatsSingleNodeResponse,
     TResult,
 )
-from glide.protobuf.command_request_pb2 import RequestType
-from glide.routes import Route
-from glide.shared.commands.core_options import (
-    FlushMode,
-    FunctionRestorePolicy,
-    InfoSection,
-)
-from glide.shared.commands.sync_commands.core import CoreCommands
+from glide.shared.routes import Route
 
 
 class ClusterCommands(CoreCommands):
