+++ python/python/glide/shared/commands/sync_commands/core.py
@@ -1,16 +1,8 @@
 # Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0
 from typing import Dict, List, Mapping, Optional, Protocol, Set, Tuple, Union, cast
 
-from glide.constants import (
-    TOK,
-    TEncodable,
-    TResult,
-    TXInfoStreamFullResponse,
-    TXInfoStreamResponse,
-)
 from glide.glide import ClusterScanCursor
 from glide.protobuf.command_request_pb2 import RequestType
-from glide.routes import Route
 from glide.shared.commands.bitmap import (
     BitFieldGet,
     BitFieldSubCommands,
@@ -59,6 +51,14 @@ from glide.shared.commands.stream import (
     StreamTrimOptions,
     _create_xpending_range_args,
 )
+from glide.shared.constants import (
+    TOK,
+    TEncodable,
+    TResult,
+    TXInfoStreamFullResponse,
+    TXInfoStreamResponse,
+)
+from glide.shared.routes import Route
 
 
 class CoreCommands(Protocol):
