+++ python/python/glide/shared/constants.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, List, Literal, Mapping, Optional, Set, TypeVar, Un
 
 from glide.protobuf.command_request_pb2 import CommandRequest
 from glide.protobuf.connection_request_pb2 import ConnectionRequest
-from glide.routes import ByAddressRoute, RandomNode, SlotIdRoute, SlotKeyRoute
+from glide.shared.routes import ByAddressRoute, RandomNode, SlotIdRoute, SlotKeyRoute
 
 OK: str = "OK"
 DEFAULT_READ_BYTES_SIZE: int = pow(2, 16)
