+++ python/python/glide/shared/routes.py
@@ -3,9 +3,9 @@
 from enum import Enum
 from typing import Optional
 
-from glide.exceptions import RequestError
 from glide.protobuf.command_request_pb2 import CommandRequest, Routes, SimpleRoutes
 from glide.protobuf.command_request_pb2 import SlotTypes as ProtoSlotTypes
+from glide.shared.exceptions import RequestError
 
 
 class SlotType(Enum):
