+++ python/tests/async_tests/test_read_from_strategy.py
@@ -5,11 +5,11 @@ from typing import Mapping, cast
 
 import pytest
 
-from glide.config import ProtocolVersion, ReadFrom
-from glide.constants import OK
 from glide.glide_client import GlideClusterClient
-from glide.routes import AllNodes, SlotIdRoute, SlotType
 from glide.shared.commands.core_options import InfoSection
+from glide.shared.config import ProtocolVersion, ReadFrom
+from glide.shared.constants import OK
+from glide.shared.routes import AllNodes, SlotIdRoute, SlotType
 from tests.async_tests.conftest import create_client
 from tests.utils.utils import get_first_result
