+++ python/tests/async_tests/test_scan.py
@@ -3,12 +3,12 @@ from typing import AsyncGenerator, List, cast
 import anyio
 import pytest
 
-from glide import ByAddressRoute
-from glide.config import ProtocolVersion
-from glide.exceptions import RequestError
 from glide.glide import ClusterScanCursor
 from glide.glide_client import GlideClient, GlideClusterClient
 from glide.shared.commands.command_args import ObjectType
+from glide.shared.config import ProtocolVersion
+from glide.shared.exceptions import RequestError
+from glide.shared.routes import ByAddressRoute
 from tests.async_tests.conftest import create_client
 from tests.utils.cluster import ValkeyCluster
 from tests.utils.utils import get_random_string
