+++ python/tests/async_tests/tests_server_modules/test_json.py
@@ -8,9 +8,6 @@ from typing import List, Optional
 
 import pytest
 
-from glide.config import ProtocolVersion
-from glide.constants import OK
-from glide.exceptions import RequestError
 from glide.glide_client import GlideClusterClient, TGlideClient
 from glide.shared.commands.batch import ClusterBatch
 from glide.shared.commands.core_options import ConditionalChange
@@ -21,6 +18,9 @@ from glide.shared.commands.server_modules.glide_json import (
     JsonArrPopOptions,
     JsonGetOptions,
 )
+from glide.shared.config import ProtocolVersion
+from glide.shared.constants import OK
+from glide.shared.exceptions import RequestError
 from tests.async_tests.test_async_client import get_random_string
