+++ python/tests/test_config.py
@@ -1,6 +1,9 @@
 # Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0
 
-from glide.config import (
+from glide.protobuf.connection_request_pb2 import ConnectionRequest
+from glide.protobuf.connection_request_pb2 import ReadFrom as ProtobufReadFrom
+from glide.protobuf.connection_request_pb2 import TlsMode
+from glide.shared.config import (
     AdvancedGlideClientConfiguration,
     AdvancedGlideClusterClientConfiguration,
     BackoffStrategy,
@@ -12,9 +15,6 @@ from glide.config import (
     PeriodicChecksStatus,
     ReadFrom,
 )
-from glide.protobuf.connection_request_pb2 import ConnectionRequest
-from glide.protobuf.connection_request_pb2 import ReadFrom as ProtobufReadFrom
-from glide.protobuf.connection_request_pb2 import TlsMode
 
 
 def test_default_client_config():
