+++ python/tests/test_proto_coded.py
@@ -4,7 +4,7 @@ import pytest
 
 from glide.protobuf.command_request_pb2 import CommandRequest, RequestType
 from glide.protobuf.response_pb2 import Response
-from glide.protobuf_codec import PartialMessageException, ProtobufCodec
+from glide.shared.protobuf_codec import PartialMessageException, ProtobufCodec
 
 
 class TestProtobufCodec:
