+++ python/tests/utils/utils.py
@@ -18,7 +18,10 @@ from typing import (
 import pytest
 from packaging import version
 
-from glide.config import (
+from glide.glide_client import GlideClient, GlideClusterClient, TGlideClient
+from glide.logger import Level as logLevel
+from glide.shared.commands.core_options import InfoSection
+from glide.shared.config import (
     AdvancedGlideClientConfiguration,
     AdvancedGlideClusterClientConfiguration,
     BackoffStrategy,
@@ -29,16 +32,13 @@ from glide.config import (
     ReadFrom,
     ServerCredentials,
 )
-from glide.constants import (
+from glide.shared.constants import (
     TClusterResponse,
     TFunctionListResponse,
     TFunctionStatsSingleNodeResponse,
     TResult,
 )
-from glide.glide_client import GlideClient, GlideClusterClient, TGlideClient
-from glide.logger import Level as logLevel
-from glide.routes import AllNodes
-from glide.shared.commands.core_options import InfoSection
+from glide.shared.routes import AllNodes
 from glide.sync import GlideClient as SyncGlideClient
 from glide.sync import GlideClusterClient as SyncGlideClusterClient
 from glide.sync import TGlideClient as TSyncGlideClient
